server/requirements.txt
absl-py==0.11.0
apt-clone==0.2.1
apturl==0.5.2
asgiref==3.3.1
astroid==2.4.2
astunparse==1.6.3
autopep8==1.5.4
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
cachetools==4.2.1
catfish==1.4.13
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.6
djangorestframework==3.12.2
docopt==0.6.2
entrypoints==0.3
Flask==1.1.2
Flask-Cors==3.0.10
flatbuffers==1.12
gast==0.3.3
google-auth==1.25.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.1.0
h5py==2.10.0
httplib2==0.14.0
idna==2.8
ifaddr==0.1.6
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lightdm-gtk-greeter-settings==1.2.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
matplotlib==3.3.4
mccabe==0.6.1
menulibre==2.2.1
mugshot==0.4.2
nemo-emblems==4.6.0
netifaces==0.10.4
nose==1.3.7
numpy==1.19.5
oauthlib==3.1.0
onboard==1.4.1
opt-einsum==3.3.0
packaging==20.3
PAM==0.4.2
path-and-address==2.0.1
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.14.0
psutil==5.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.6.0
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
Pygments==2.3.1
PyGObject==3.36.0
PyICU==2.4.2
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyparted==3.11.2
pyperclip==1.8.1
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-xapp==2.0.1
python-xlib==0.23
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7
SecretStorage==2.3.1
selenium==3.141.0
setproctitle==1.1.10
sgt-launcher==0.2.5
simplejson==3.16.0
six==1.15.0
soupsieve==1.9.5
sqlparse==0.4.1
systemd-python==234
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
tinycss2==1.0.2
toml==0.10.2
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
Unidecode==1.1.1
urllib3==1.25.8
wadllib==1.3.3
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
xcffib==0.8.1
xkit==0.0.0
zeroconf==0.24.4
